+++ SolastaUnfinishedBusiness/Displays/RulesDisplay.cs
@@ -508,10 +508,10 @@ internal static class RulesDisplay
 
         using (UI.HorizontalScope())
         {
-            UI.Label(Gui.Localize("Caption/&TargetFilteringAllyCreature"), UI.Width((float)100));
+            UI.Label(Gui.Localize("Caption/&TargetFilteringAllyCreature"), UI.Width(100f));
 
             intValue = Main.Settings.CriticalHitModeAllies;
-            if (UI.SelectionGrid(ref intValue, Options, Options.Length, 4, UI.Width((float)220)))
+            if (UI.SelectionGrid(ref intValue, Options, Options.Length, 4, UI.Width(220f)))
             {
                 Main.Settings.CriticalHitModeAllies = intValue;
             }
@@ -519,10 +519,10 @@ internal static class RulesDisplay
 
         using (UI.HorizontalScope())
         {
-            UI.Label(Gui.Localize("Caption/&TargetFilteringEnemyCreature"), UI.Width((float)100));
+            UI.Label(Gui.Localize("Caption/&TargetFilteringEnemyCreature"), UI.Width(100f));
 
             intValue = Main.Settings.CriticalHitModeEnemies;
-            if (UI.SelectionGrid(ref intValue, Options, Options.Length, 4, UI.Width((float)220)))
+            if (UI.SelectionGrid(ref intValue, Options, Options.Length, 4, UI.Width(220f)))
             {
                 Main.Settings.CriticalHitModeEnemies = intValue;
             }
@@ -530,10 +530,10 @@ internal static class RulesDisplay
 
         using (UI.HorizontalScope())
         {
-            UI.Label(Gui.Localize("Action/&NeutralCreatureTitle"), UI.Width((float)100));
+            UI.Label(Gui.Localize("Action/&NeutralCreatureTitle"), UI.Width(100f));
 
             intValue = Main.Settings.CriticalHitModeNeutral;
-            if (UI.SelectionGrid(ref intValue, Options, Options.Length, 4, UI.Width((float)220)))
+            if (UI.SelectionGrid(ref intValue, Options, Options.Length, 4, UI.Width(220f)))
             {
                 Main.Settings.CriticalHitModeNeutral = intValue;
             }
