+++ SolastaUnfinishedBusiness/Displays/TranslationsDisplay.cs
@@ -16,7 +16,7 @@ internal static class TranslationsDisplay
 
         using (UI.HorizontalScope())
         {
-            UI.Label(Gui.Localize("ModUi/&TargetLanguage"), UI.Width((float)120));
+            UI.Label(Gui.Localize("ModUi/&TargetLanguage"), UI.Width(120f));
 
             var intValue = Array.IndexOf(TranslatorContext.AvailableLanguages, Main.Settings.SelectedLanguageCode);
 
@@ -24,7 +24,7 @@ internal static class TranslationsDisplay
                     ref intValue,
                     TranslatorContext.AvailableLanguages,
                     TranslatorContext.AvailableLanguages.Length,
-                    3, UI.Width((float)300)))
+                    3, UI.Width(300f)))
             {
                 Main.Settings.SelectedLanguageCode = TranslatorContext.AvailableLanguages[intValue];
             }
@@ -47,8 +47,8 @@ internal static class TranslationsDisplay
 
                 UI.Label(
                     userCampaign.Author.Substring(0, Math.Min(16, userCampaign.Author.Length)).Bold().Orange(),
-                    UI.Width((float)120));
-                UI.Label(userCampaign.Title.Bold().Italic(), UI.Width((float)300));
+                    UI.Width(120f));
+                UI.Label(userCampaign.Title.Bold().Italic(), UI.Width(300f));
 
                 if (TranslatorContext.TranslatorBehaviour.CurrentExports.TryGetValue(exportName, out var status))
                 {
@@ -72,7 +72,7 @@ internal static class TranslationsDisplay
                             TranslatorContext.TranslatorBehaviour.Cancel(userCampaign.Title);
                         }
                     },
-                    UI.Width((float)200));
+                    UI.Width(200f));
             }
         }
